implemented the collision logic for the game. When a player collides with the sides of a platform or with a deadly object like a spike they die and respawn. Players can also go through no eadly objects like chains.
Bugs:
Would not detect the deadly object because of a calculation error, would not see the adjacent tile correctly

Colliding with the side of a platform would not kill the player because I was setting its x velocity would be 0 before the check happened causing the error.

In generate tiles, the tile id is subtracted by 1 so the ids were not matching up and no collisoin was happening.