Added the nature objects to the tile class. They're then used in the map to know that the player can pass through them and not collide.
Coin is now a collectable. When the player collides with it it is added to the players property and removed from the map. This will be used for a player score later on.

To put the coins back after collecting them I keep track of the map definition inside of map, then playstate will call a reset method that adds all the tiles back