Integrated map 2 similarly to ,map one. Added the tile ids to tile and then map ues them to handle collisions with the collision detector.
New feature: added a jump pad where when the use touches it they boost up into the air using y velocity.

Map selection feature: We first load all the levels in main and pass it to the playstate. That is where the map is created. When the user selects the level,  the enter method in playstate creates that map and then makes all states use that map.

Bug: We were still pointing to the old map making the collision detect the wrong tiles. So I manually changed it in playstate.enter. i knew this because I had noticed that the player would die in the exact same spot as where a spike was in map one. By far my biggest most time consuming bug.